Split python unit test ASM config parser code out to an independent file
This will be needed once we have multiple file type configs getting
parsed
